Try and make the watch images integration test less flaky
The watch-images Cypress test has been very flaky in CI, especially
recently. This commit tries to make it more reliable by using the
Cypress `wait()` feature [[1]]. By waiting for Browsersync to reload the
page after the image has been copied, we can be sure of when the image
has been copied to the public folder.

Note that Browsersync doesn't put its reloading script on our 404 pages
or on images, so this commit has to add a page fixture for displaying
the image.

[1]: https://docs.cypress.io/api/commands/wait