Set Cypress timeouts for CI only
We want Cypress to have long timeouts and do lots of retries when
running tests on CI, but when running tests locally it is a little
annoying. This commit moves the definition of the config variables to
environment variables in the CI pipeline, rather than in the global
configuration file. See Cypress docs [[1]] for details on how this
works.

[1]: https://docs.cypress.io/guides/guides/environment-variables#Option-3-CYPRESS_